Computer -Vision-Face-Analyzer/requirements.txt
azure-ai-vision-face==1.0.0b2
azure-core
pillow
python-dotenv
